change class name.  code  clear up
